// src/bot/types/vaccination-coverage.interface.ts
export interface VaccinationCoverage {
  coddepto: string;
  departamento: string;
  a_o: string;
  biol_gico: string;
  cobertura_de_vacunaci_n: string;
  indicator1?: string;
  indicator1_1?: string;
  indicator1_2?: string;
  indicator1_3?: string;
  indicator1_4?: string;
  indicator1_5?: string;
}